+++ package.json
@@ -42,6 +42,7 @@
     "react-redux": "7.2.0",
     "react-router-dom": "5.2.0",
     "redux": "4.0.5",
-    "redux-thunk": "2.3.0"
+    "redux-thunk": "2.3.0",
+    "reselect": "4.0.0"
   }
 }
