+++ src/reducer/data/selectors.js
@@ -1,5 +1,8 @@
 import NameSpace from "../name-space";
+import {createSelector} from "reselect";
+import {DEFAULT_GENRE} from "../film/film";
 
+const MAX_GENRES = 10;
 const NAME_SPACE = NameSpace.DATA;
 
 export const getAllFilms = (state) => {
@@ -9,3 +12,20 @@ export const getAllFilms = (state) => {
 export const getPromoFilm = (state) => {
   return state[NAME_SPACE].promoFilm;
 };
+
+export const getFilters = createSelector(
+    getAllFilms,
+    (films) => {
+      const filters = [DEFAULT_GENRE];
+      films.forEach((film) => {
+        const filmGenres = film.genre.split(`, `);
+        filmGenres.forEach((genre) => {
+          if (filters.indexOf(genre) === -1) {
+            filters.push(genre);
+          }
+        });
+      });
+
+      return filters.slice(0, MAX_GENRES);
+    }
+);
